test: split functional test code
